+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features
 
 - Add option to attach raw tombstone protobuf on native crash events ([#5446](https://github.com/getsentry/sentry-java/pull/5446))
-  - Enable via `options.isAttachTombstone = true` or manifest: `<meta-data android:name="io.sentry.tombstone.attach" android:value="true" />`
+  - Enable via `options.isAttachRawTombstone = true` or manifest: `<meta-data android:name="io.sentry.tombstone.attach-raw" android:value="true" />`
 - Add support to configure reporting historical ANRs via `AndroidManifest.xml` using the  `io.sentry.anr.report-historical` attribute ([#5387](https://github.com/getsentry/sentry-java/pull/5387))
 
 ### Dependencies